Part 1 of Learning Numpy
This is my first part in the series of learning NumPy.

This part is based on the documentation and the Quickstart guide.

All rights are reserved to the SciPy community - where of Links and comments are part of the Document.

As it currently stands - I am looking to implement a Indexing and a Post Mortem of this part - As i wish to include a Systematic Fashion of Deduction in processing.